[MS] Moved all html handling into a jsontohtml package, extracted methods into a utility file
