+++ jsontohtml/parser.go
@@ -1,4 +1,4 @@
-package parser
+package jsontohtml
 
 import (
 	"encoding/json"
@@ -9,7 +9,6 @@ import (
 	"strings"
 
 	"github.com/ONSdigital/dp-table-renderer/models"
-	"github.com/ONSdigital/dp-table-renderer/renderer"
 	"github.com/go-ns/log"
 	"golang.org/x/net/html"
 	"golang.org/x/net/html/atom"
@@ -43,7 +42,7 @@ func ParseHTML(request *models.ParseRequest) ([]byte, error) {
 	// todo parse table to requestJSON
 	findNode(sourceTable, atom.Tbody)
 
-	previewHTML, err := renderer.RenderHTML(requestJSON)
+	previewHTML, err := RenderHTML(requestJSON)
 	if err != nil {
 		log.Error(err, log.Data{"message": "Unable to render preview HTML", "ParseRequest": request, "RenderRequest": requestJSON})
 		return nil, err
@@ -83,60 +82,4 @@ func marshalResponse(response ResponseModel) ([]byte, error) {
 		err = writer.Flush()
 	}
 	return b.Bytes(), err
-}
-
-// find an attribute for the node - returns empty string if not found
-func getAttribute(node *html.Node, key string) string {
-	for _, attr := range node.Attr {
-		if attr.Key == key {
-			return attr.Val
-		}
-	}
-	return ""
-}
-
-// depth-first search for the first node of the given type
-func findNode(n *html.Node, a atom.Atom) *html.Node {
-	return findNodeWithAttributes(n, a, nil)
-}
-
-// depth-first search for the first node of the given type with the given attributes
-func findNodeWithAttributes(n *html.Node, a atom.Atom, attr map[string]string) *html.Node {
-	for c := n.FirstChild; c != nil; c = c.NextSibling {
-		if c.DataAtom == a && hasAttributes(c, attr) {
-			return c
-		}
-		gc := findNodeWithAttributes(c, a, attr)
-		if gc != nil {
-			return gc
-		}
-	}
-	return nil
-}
-
-// return true if the given node has all the attribute values
-func hasAttributes(n *html.Node, attr map[string]string) bool {
-	for key, value := range attr {
-		if getAttribute(n, key) != value {
-			return false
-		}
-	}
-	return true
-}
-
-// returns all child nodes of the given type
-func findNodes(n *html.Node, a atom.Atom) []*html.Node {
-	return findNodesWithAttributes(n, a, nil)
-}
-
-// returns all child nodes of the given type with the given attributes
-func findNodesWithAttributes(n *html.Node, a atom.Atom, attr map[string]string) []*html.Node {
-	var result []*html.Node
-	for c := n.FirstChild; c != nil; c = c.NextSibling {
-		if c.DataAtom == a && hasAttributes(c, attr) {
-			result = append(result, c)
-		}
-		result = append(result, findNodesWithAttributes(c, a, attr)...)
-	}
-	return result
 }
